refactor: split out DonateBrowseButtons for reuse
